+++ src/command_modules/azure-cli-interactive/azure/cli/command_modules/interactive/__init__.py
@@ -3,11 +3,12 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-# pylint: disable=line-too-long
+import azure.cli.command_modules.interactive._help  # pylint: disable=unused-import
 
-from azure.cli.core.commands import register_cli_argument
 
-from azclishell.color_styles import get_options as style_options
+def load_params(_):
+    import azure.cli.command_modules.interactive._params  # pylint: disable=redefined-outer-name
 
-register_cli_argument('shell', 'style', options_list=('--style', '-s'),
-                      help='The colors of the shell.', choices=style_options())
+
+def load_commands():
+    import azure.cli.command_modules.interactive.commands  # pylint: disable=redefined-outer-name