+++ src/command_modules/azure-cli-interactive/azure/cli/command_modules/interactive/_help.py
@@ -3,12 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-import azure.cli.command_modules.shell._help  # pylint: disable=unused-import
+from azure.cli.core.help_files import helps
 
-
-def load_params(_):
-    import azure.cli.command_modules.shell._params  # pylint: disable=redefined-outer-name
-
-
-def load_commands():
-    import azure.cli.command_modules.shell.commands  # pylint: disable=redefined-outer-name
+helps['launch-interactive'] = """
+            type: command
+            short-summary: Start the interactive shell for Microsoft Azure CLI (preview)
+            """