+++ src/command_modules/azure-cli-interactive/azure/cli/command_modules/interactive/commands.py
@@ -5,4 +5,5 @@
 
 from azure.cli.core.commands import cli_command
 
-cli_command(__name__, 'shell', 'azure.cli.command_modules.shell.custom#start_shell')
+cli_command(__name__,
+            'launch-interactive', 'azure.cli.command_modules.interactive.custom#start_shell')